requirements.txt
PyQt6>=6.0.0
pyperclip>=1.8.2
pynput>=1.7.6
pillow>=9.0.0
keyboard>=0.13.5
pywin32>=303

